requirements.txt
streamlit
openai
python-dotenv
Pillow
torch
torchvision

